Create a module and import and export function
