+++ src/index.js
@@ -1,7 +1,8 @@
 import { join } from 'lodash';
 import './style.css';
+import { check } from './modules/check';
 
-const tasks = [
+export const tasks = [
   {
     description: 'Wash dishes',
     completed: false,
@@ -37,19 +38,4 @@ const displayTasks = () => {
 };
 
 displayTasks();
-
-
-const checkboxed = document.getElementsByClassName('checkboxed');
-for (let j = 0; j < checkboxed.length; j++) {
-  checkboxed[j].addEventListener('change', () => {
-    if(checkboxed[j].checked) {
-      checkboxed[j].nextElementSibling.style.textDecoration = 'line-through';
-      tasks[j].completed = true;
-      console.log(tasks[j].completed);
-    } else {
-      checkboxed[j].nextElementSibling.style.textDecoration = 'none';
-      tasks[j].completed = false;
-      console.log(tasks[j].completed);
-    }
-  });
-}
+check();
